Consolidate v1 and v2 metric parsers
