+++ plugins/inputs/prometheus/prometheus.go
@@ -27,7 +27,7 @@ import (
 	"github.com/influxdata/telegraf/internal"
 	httpconfig "github.com/influxdata/telegraf/plugins/common/http"
 	"github.com/influxdata/telegraf/plugins/inputs"
-	parserV2 "github.com/influxdata/telegraf/plugins/parsers/prometheus"
+	parser "github.com/influxdata/telegraf/plugins/parsers/prometheus"
 )
 
 //go:embed sample.conf
@@ -207,6 +207,10 @@ func (p *Prometheus) Init() error {
 		return err
 	}
 
+	if p.MetricVersion == 0 {
+		p.MetricVersion = 1
+	}
+
 	ctx := context.Background()
 	if p.ResponseTimeout != 0 {
 		p.HTTPClientConfig.Timeout = p.ResponseTimeout
@@ -357,14 +361,14 @@ func (p *Prometheus) Gather(acc telegraf.Accumulator) error {
 
 func (p *Prometheus) gatherURL(u URLAndAddress, acc telegraf.Accumulator) error {
 	var req *http.Request
-	var err error
 	var uClient *http.Client
-	var metrics []telegraf.Metric
 	if u.URL.Scheme == "unix" {
 		path := u.URL.Query().Get("path")
 		if path == "" {
 			path = "/metrics"
 		}
+
+		var err error
 		addr := "http://localhost" + path
 		req, err = http.NewRequest("GET", addr, nil)
 		if err != nil {
@@ -390,6 +394,7 @@ func (p *Prometheus) gatherURL(u URLAndAddress, acc telegraf.Accumulator) error 
 		if u.URL.Path == "" {
 			u.URL.Path = "/metrics"
 		}
+		var err error
 		req, err = http.NewRequest("GET", u.URL.String(), nil)
 		if err != nil {
 			return fmt.Errorf("unable to create new request %q: %w", u.URL.String(), err)
@@ -414,6 +419,7 @@ func (p *Prometheus) gatherURL(u URLAndAddress, acc telegraf.Accumulator) error 
 		req.Header.Set(key, value)
 	}
 
+	var err error
 	var resp *http.Response
 	if u.URL.Scheme != "unix" {
 		//nolint:bodyclose // False positive (because of if-else) - body will be closed in `defer`
@@ -436,16 +442,13 @@ func (p *Prometheus) gatherURL(u URLAndAddress, acc telegraf.Accumulator) error 
 		return fmt.Errorf("error reading body: %w", err)
 	}
 
-	if p.MetricVersion == 2 {
-		parser := parserV2.Parser{
-			Header:          resp.Header,
-			IgnoreTimestamp: p.IgnoreTimestamp,
-		}
-		metrics, err = parser.Parse(body)
-	} else {
-		metrics, err = Parse(body, resp.Header, p.IgnoreTimestamp)
+	// Parse the metrics
+	parser := parser.Parser{
+		Header:          resp.Header,
+		MetricVersion:   p.MetricVersion,
+		IgnoreTimestamp: p.IgnoreTimestamp,
 	}
-
+	metrics, err := parser.Parse(body)
 	if err != nil {
 		return fmt.Errorf("error reading metrics for %q: %w", u.URL, err)
 	}
