+++ plugins/parsers/prometheus/parser_v1.go
@@ -17,10 +17,9 @@ import (
 
 	"github.com/influxdata/telegraf"
 	"github.com/influxdata/telegraf/metric"
-	"github.com/influxdata/telegraf/plugins/parsers/prometheus"
 )
 
-func Parse(buf []byte, header http.Header, ignoreTimestamp bool) ([]telegraf.Metric, error) {
+func (p *Parser) parse_v1(buf []byte) ([]telegraf.Metric, error) {
 	var parser expfmt.TextParser
 	var metrics []telegraf.Metric
 	var err error
@@ -33,7 +32,7 @@ func Parse(buf []byte, header http.Header, ignoreTimestamp bool) ([]telegraf.Met
 	// Prepare output
 	metricFamilies := make(map[string]*dto.MetricFamily)
 
-	if isProtobuf(header) {
+	if isProtobuf(p.Header) {
 		for {
 			mf := &dto.MetricFamily{}
 			if _, ierr := pbutil.ReadDelimited(reader, mf); ierr != nil {
@@ -56,35 +55,35 @@ func Parse(buf []byte, header http.Header, ignoreTimestamp bool) ([]telegraf.Met
 	for metricName, mf := range metricFamilies {
 		for _, m := range mf.Metric {
 			// reading tags
-			tags := prometheus.GetTagsFromLabels(m, nil)
+			tags := GetTagsFromLabels(m, nil)
 
 			// reading fields
 			var fields map[string]interface{}
 			if mf.GetType() == dto.MetricType_SUMMARY {
 				// summary metric
-				fields = makeQuantiles(m)
+				fields = makeQuantilesV1(m)
 				fields["count"] = float64(m.GetSummary().GetSampleCount())
 				//nolint:unconvert // Conversion may be needed for float64 https://github.com/mdempsky/unconvert/issues/40
 				fields["sum"] = float64(m.GetSummary().GetSampleSum())
 			} else if mf.GetType() == dto.MetricType_HISTOGRAM {
 				// histogram metric
-				fields = makeBuckets(m)
+				fields = makeBucketsV1(m)
 				fields["count"] = float64(m.GetHistogram().GetSampleCount())
 				//nolint:unconvert // Conversion may be needed for float64 https://github.com/mdempsky/unconvert/issues/40
 				fields["sum"] = float64(m.GetHistogram().GetSampleSum())
 			} else {
 				// standard metric
-				fields = getNameAndValue(m)
+				fields = getNameAndValueV1(m)
 			}
 			// converting to telegraf metric
 			if len(fields) > 0 {
 				var t time.Time
-				if !ignoreTimestamp && m.TimestampMs != nil && *m.TimestampMs > 0 {
+				if !p.IgnoreTimestamp && m.TimestampMs != nil && *m.TimestampMs > 0 {
 					t = time.Unix(0, *m.TimestampMs*1000000)
 				} else {
 					t = now
 				}
-				m := metric.New(metricName, tags, fields, t, prometheus.ValueType(mf.GetType()))
+				m := metric.New(metricName, tags, fields, t, ValueType(mf.GetType()))
 				metrics = append(metrics, m)
 			}
 		}
@@ -105,7 +104,7 @@ func isProtobuf(header http.Header) bool {
 }
 
 // Get Quantiles from summary metric
-func makeQuantiles(m *dto.Metric) map[string]interface{} {
+func makeQuantilesV1(m *dto.Metric) map[string]interface{} {
 	fields := make(map[string]interface{})
 	for _, q := range m.GetSummary().Quantile {
 		if !math.IsNaN(q.GetValue()) {
@@ -117,7 +116,7 @@ func makeQuantiles(m *dto.Metric) map[string]interface{} {
 }
 
 // Get Buckets  from histogram metric
-func makeBuckets(m *dto.Metric) map[string]interface{} {
+func makeBucketsV1(m *dto.Metric) map[string]interface{} {
 	fields := make(map[string]interface{})
 	for _, b := range m.GetHistogram().Bucket {
 		fields[fmt.Sprint(b.GetUpperBound())] = float64(b.GetCumulativeCount())
@@ -126,7 +125,7 @@ func makeBuckets(m *dto.Metric) map[string]interface{} {
 }
 
 // Get name and value from metric
-func getNameAndValue(m *dto.Metric) map[string]interface{} {
+func getNameAndValueV1(m *dto.Metric) map[string]interface{} {
 	fields := make(map[string]interface{})
 	if m.Gauge != nil {
 		if !math.IsNaN(m.GetGauge().GetValue()) {
